# 🐇 Avoid making two copies of SyntaxToken at each step of the formatting operations chain