--- a/src/Features/CSharp/Portable/ChangeSignature/ChangeSignatureFormattingRule.cs
+++ b/src/Features/CSharp/Portable/ChangeSignature/ChangeSignatureFormattingRule.cs
@@ -45,14 +45,14 @@ namespace Microsoft.CodeAnalysis.CSharp.ChangeSignature
             }
         }
 
-        public override AdjustNewLinesOperation GetAdjustNewLinesOperation(SyntaxToken previousToken, SyntaxToken currentToken, in NextGetAdjustNewLinesOperation nextOperation)
+        public override AdjustNewLinesOperation GetAdjustNewLinesOperation(in SyntaxToken previousToken, in SyntaxToken currentToken, in NextGetAdjustNewLinesOperation nextOperation)
         {
             if (previousToken.Kind() == SyntaxKind.CommaToken && s_allowableKinds.Contains(previousToken.Parent.Kind()))
             {
                 return FormattingOperations.CreateAdjustNewLinesOperation(0, AdjustNewLinesOption.PreserveLines);
             }
 
-            return base.GetAdjustNewLinesOperation(previousToken, currentToken, in nextOperation);
+            return base.GetAdjustNewLinesOperation(in previousToken, in currentToken, in nextOperation);
         }
     }
 }