--- a/src/Features/Core/Portable/MetadataAsSource/AbstractMetadataAsSourceService.AbstractMetadataFormattingRule.cs
+++ b/src/Features/Core/Portable/MetadataAsSource/AbstractMetadataAsSourceService.AbstractMetadataFormattingRule.cs
@@ -17,11 +17,11 @@ namespace Microsoft.CodeAnalysis.MetadataAsSource
             protected abstract bool IsNewLine(char c);
 
             public override AdjustNewLinesOperation GetAdjustNewLinesOperation(
-                    SyntaxToken previousToken, SyntaxToken currentToken, in NextGetAdjustNewLinesOperation nextOperation)
+                    in SyntaxToken previousToken, in SyntaxToken currentToken, in NextGetAdjustNewLinesOperation nextOperation)
             {
                 if (previousToken.RawKind == 0 || currentToken.RawKind == 0)
                 {
-                    nextOperation.Invoke();
+                    nextOperation.Invoke(in previousToken, in currentToken);
                 }
 
                 var betweenMembersAndUsingsOperation = GetAdjustNewLinesOperationBetweenMembersAndUsings(previousToken, currentToken);
@@ -30,7 +30,7 @@ namespace Microsoft.CodeAnalysis.MetadataAsSource
                     return betweenMembersAndUsingsOperation;
                 }
 
-                return nextOperation.Invoke();
+                return nextOperation.Invoke(in previousToken, in currentToken);
             }
 
             protected int GetNumberOfLines(IEnumerable<SyntaxTrivia> triviaList)